Copy server's IP to client
I'd love some help with this - this solution feels icky. I couldn't
find a way to get Vagrant to tell me what a guest's IP address is, and
solutions using libvirt involved scrobbling around inside libvirt's
guts to figure it out. I went with a brute-force approach where the
server's playbook writes the IP address to a file and the client
playbook picks up that file and uses the contents as the address. Yuk!

The implementation adds two small roles: clientws and epicws. epicws
writes the file and clientws waits for it to show up and then reads it
into the ws_ip variable.

Signed-off-by: Toby Cabot <toby@acnodal.io>