Use a patch to set the web service's name/address binding
We used to have static IP addresses so we could build a "debug"
version of the puregw manifest, but now we're using the Vagrant
management NIC which gets its address from DHCP.

This change uses a kubectl patch to add the hostAliases patch to the
puregw deployment and daemonset.

Getting the quoting correct in gateway.yml was not fun.

Signed-off-by: Toby Cabot <toby@acnodal.io>